--- a/debian/changelog
+++ b/debian/changelog
@@ -1,5 +1,6 @@
   [Michele Simionato]
-  * Removed the syntax for magnitude-dependent magnitude distance, since
+  * Enabled magnitude-dependent pointsource_distance and maximum_distance
+  * Removed the syntax for magnitude-dependent maximum distance, since
     now it can be automatically determined by the engine
   * Saving more information in the case of single-site classical hazard
   * Extended `pointsource_distance` to generic sources