--- a/demos/hazard/AreaSourceClassicalPSHA/job.ini
+++ b/demos/hazard/AreaSourceClassicalPSHA/job.ini
@@ -43,6 +43,7 @@ intensity_measure_types_and_levels = {
   "SA(2.0)": logscale(0.005, 2.13, 20)}
 truncation_level = 3
 maximum_distance = 200.0
+pointsource_distance = 100.0
 
 [output]