--- a/openquake/baselib/workerpool.py
+++ b/openquake/baselib/workerpool.py
@@ -234,7 +234,7 @@ class WorkerPool(object):
                 elif cmd == 'get_num_workers':
                     ctrlsock.send(self.num_workers)
                 elif cmd == 'get_executing':
-                    ctrlsock.send(' '.join(os.listdir(self.executing)))
+                    ctrlsock.send(' '.join(sorted(os.listdir(self.executing))))
         shutil.rmtree(self.executing)
 
     def stop(self):