--- a/openquake/commonlib/oqvalidation.py
+++ b/openquake/commonlib/oqvalidation.py
@@ -177,7 +177,7 @@ class OqParam(valid.ParamSet):
     specific_assets = valid.Param(valid.namelist, [])
     split_by_magnitude = valid.Param(valid.boolean, False)
     task_duration = valid.Param(valid.positiveint, None)  # used in ebrisk
-    task_multiplier = valid.Param(valid.positiveint, 10)  # used in classical
+    task_multiplier = valid.Param(valid.positiveint, 20)  # used in classical
     taxonomies_from_model = valid.Param(valid.boolean, False)
     time_event = valid.Param(str, None)
     truncation_level = valid.Param(valid.NoneOr(valid.positivefloat), None)