… doc/adv-manual/common-mistakes.rst
… openquake/calculators/tests/__init__.py
… openquake/calculators/tests/classical_test.py
… openquake/hazardlib/site.py
… openquake/hazardlib/tests/contexts_test.py
… openquake/qa_tests_data/classical/case_19/expected/hazard_curve-mean_PGA.csv
… openquake/qa_tests_data/classical/case_19/expected/hazard_curve-mean_SA(0.1).csv
… openquake/qa_tests_data/classical/case_19/expected/hazard_curve-mean_SA(0.15).csv
… openquake/qa_tests_data/classical/case_19/job.ini
… openquake/qa_tests_data/classical/case_36/job.ini
… openquake/qa_tests_data/classical/case_39/job.ini
… openquake/qa_tests_data/classical/case_43/job.ini
… openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.02-PGA-sid-0_Src.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-PGA-sid-0_Src.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/poe-0.1-SA(0.025)-sid-0_Src.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-0_Lon_Lat.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-0_Mag.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-0_Mag_Dist.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-1_Lon_Lat.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-1_Mag.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-PGA-sid-0-poe-1_Mag_Dist.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-0_Lon_Lat.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-0_Mag.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-0_Mag_Dist.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-1_Lon_Lat.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-1_Mag.csv
… openquake/qa_tests_data/disagg/case_1/expected_output/rlz-0-SA(0.025)-sid-0-poe-1_Mag_Dist.csv
… openquake/qa_tests_data/disagg/case_6/job.ini
… openquake/qa_tests_data/event_based/case_19/job_grid.ini
… openquake/qa_tests_data/event_based_risk/case_4/job.ini