Now it can be a dictionary TRT -> distance and it enables the magnitude-dependent maximum distance feature. The speedup is a factor of 2 for a Colombia calculation with 1077 sites:
```
  master
   calc_33573                   time_sec memory_mb counts   
   ============================ ======== ========= =========
   total classical              123,667  92        347      
   computing mean_std           81,562   0.0       6,688,037
   get_poes                     33,452   0.0       6,688,037
   total classical_split_filter 13,838   110       101      
   composing pnes               11,493   0.0       6,688,037
   ClassicalCalculator.run      7,385    1,082     1        
   make_contexts                6,574    0.0       2,766,229
   iter_ruptures                2,930    0.0       158,121  

   this branch
   calc_33574                   time_sec memory_mb counts   
   ============================ ======== ========= =========
   total classical              61,461   88        347      
   computing mean_std           31,232   0.0       2,244,391
   get_poes                     22,724   0.0       2,244,391
   composing pnes               8,155    0.0       2,244,391
   total classical_split_filter 7,906    120       101      
   iter_ruptures                3,096    0.0       158,121  
   ClassicalCalculator.run      2,918    1,098     1        
   make_contexts                2,750    0.0       1,807,793
```
The hazard curves are the same except for very small probabilities (<1E-3).
![large-probs](https://user-images.githubusercontent.com/2463856/68924996-50be1980-0782-11ea-805d-6e5f6fde4ba8.png)
![small-probs](https://user-images.githubusercontent.com/2463856/68925005-574c9100-0782-11ea-97cb-c2501e0cebc3.png)